+++ src/extensions/rules/rulesext.h
@@ -137,4 +137,20 @@ public:
      *  to allow customizing AI difficulties without affecting the human player.
      */
     DifficultyClass DiffHuman;
+
+    /**
+     *  Sounds played when a unit is promoted.
+     */
+    VocType UpgradeVeteranSound;
+    VocType UpgradeEliteSound;
+
+    /**
+     *  EVA announcement when a unit is promoted.
+     */
+    VoxType VoxUnitPromoted;
+
+    /**
+     *  The number of frames that a newly elite unit will flash for.
+     */
+    int EliteFlashTimer;
 };