+++ src/extensions/techno/technoext.cpp
@@ -55,7 +55,8 @@ TechnoClassExtension::TechnoClassExtension(const TechnoClass *this_ptr) :
     ElectricBolt(nullptr),
     Storage(Tiberiums.Count()),
     SpawnManager(nullptr),
-    SpawnOwner(nullptr)
+    SpawnOwner(nullptr),
+    LastVeterancy(VETERANCY_NONE)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("TechnoClassExtension::TechnoClassExtension - Name: %s (0x%08X)\n", Name(), (uintptr_t)(This()));